backend/requirements.txt
# Google ADK and AI dependencies
google-adk>=0.1.0
google-generativeai>=0.8.0

# FastAPI and web server dependencies  
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Additional dependencies for financial crime monitoring
python-dateutil>=2.8.0
python-multipart>=0.0.6
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
